requirements.txt
requests
openai
python-dotenv
googlesearch-python
PyYAML
beautifulsoup4
